[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 1112 KB, Score: 100 point -BaekjoonHub

+++ C99/백준/Bronze/15964. 이상한 기호/이상한 기호.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
-int p(int z1,int z2){
-    int rsl=0;
+long long p(int z1,int z2){
+    long long rsl=0;
     rsl = z1+z2;
     return rsl;
 }
@@ -13,5 +13,5 @@ int m(int z1, int z2){
 int main(void){
     int A=0,B=0;
     scanf("%d %d", &A, &B);
-    printf("%d", p(A,B) * m(A,B));
+    printf("%lld", p(A,B) * m(A,B));
 }
